Adding PowerShell-based scripts for Linux and macOS
