+++ .travis.yml
@@ -1,50 +1,38 @@
 language: cpp
 
 os:
-  - osx
+  # - osx
   - linux
 
 compiler:
-  - clang
+  # - clang
   - gcc
 
-env:
-  matrix:
-    - CMAKE_MINOR=3.5 CMAKE_VER=3.5.0
-    - CMAKE_MINOR=3.7 CMAKE_VER=3.7.0
-    - CMAKE_MINOR=3.7 CMAKE_VER=3.7.2
+dist: trusty
+sudo: false
+addons:
+  apt:
+    sources:
+      - sourceline: deb [arch=amd64] https://packages.microsoft.com/ubuntu/14.04/prod trusty main
+        key_url: https://packages.microsoft.com/keys/microsoft.asc
+    packages:
+      - powershell
 
+env:
+  - CMAKE_MINOR=3.7 CMAKE_VER=3.7.2
 
 before_install:
   - |
     if [ $TRAVIS_OS_NAME == "linux" ]; then
       export DISPLAY=:99.0
       sh -e /etc/init.d/xvfb start
-      curl https://cmake.org/files/v${CMAKE_MINOR}/cmake-${CMAKE_VER}-Linux-x86_64.sh -o /tmp/cmake.sh
-      sudo sh /tmp/cmake.sh --exclude-subdir --prefix=/usr/local
       sleep 3
     else
-      curl https://cmake.org/files/v${CMAKE_MINOR}/cmake-${CMAKE_VER}-Darwin-x86_64.tar.gz -o /tmp/cmake.tgz
-      pushd /tmp
-      tar xf /tmp/cmake.tgz
-      if test -e /Applications/CMake.app; then
-        sudo rm -rf /Applications/CMake.app
-      fi
-      sudo cp -r cmake-${CMAKE_VER}-Darwin-x86_64/CMake.app /Applications/
-      sudo /Applications/CMake.app/Contents/bin/cmake-gui --install
-      popd
+      brew tap caskroom/cask       || exit 2
+      brew cask install powershell || exit 2
     fi
 
-install:
-  - npm install
-  - npm run vscode:prepublish
-  - bash scripts/prepare_test.sh
-
-script:
-  - node --version
-  - cmake --version
-  - bash scripts/run_tests.sh
-  - bash scripts/run_extension_tests.sh
+script: pwsh -NonInteractive -NoProfile -NoLogo scripts/ci.ps1
 
 notifications:
   email:
