+++ .vscode/tasks.json
@@ -1,3 +1,13 @@
 {
-    "version": "2.0.0"
+    "version": "2.0.0",
+    "tasks": [
+        {
+            "type": "npm",
+            "script": "compile",
+            "problemMatcher": [
+                "$tsc-watch"
+            ],
+            "isBackground": true
+        }
+    ]
 }
