+++ test/extension-tests/successful-build/project-folder/.vscode/settings.json
@@ -1,4 +1,3 @@
 {
-    "cmake.buildDirectory": "${workspaceRoot}/build",
-    "cmake.preferredGenerators": [ ]
+    "cmake.buildDirectory": "${workspaceRoot}/build"
 }
