+++ test/extension-tests/vs-preferred-gen/project-folder/.vscode/settings.json
@@ -1,0 +1,4 @@
+{
+    "cmake.buildDirectory": "${workspaceRoot}/build",
+    "cmake.preferredGenerators": [ ]
+}
