+++ test/extension-tests/vs-preferred-gen/project-folder/CMakeLists.txt
